Fix 'cast from pointer to integer of different size' warning in Win64 build
The "long" data type on 64-bit Windows isn't 64-bit, so isn't enough to hold
a pointer. So use intptr_t instead.

This warning was introduced by commit a983aed4 (Do an SSL rehandshake
when the server sends "X-CSTP-Rekey-Method: ssl".)

+++ gnutls.c
@@ -1973,7 +1973,7 @@ int cstp_handshake(struct openconnect_info *vpninfo, unsigned init)
 	int err;
 	int ssl_sock = -1;
 
-	ssl_sock = (long)gnutls_transport_get_ptr(vpninfo->https_sess);
+	ssl_sock = (intptr_t)gnutls_transport_get_ptr(vpninfo->https_sess);
 
 	while ((err = gnutls_handshake(vpninfo->https_sess))) {
 		if (err == GNUTLS_E_AGAIN) {
